# Improve performance by inlining methods and avoiding loops