This provides a ~4× performance improvement in both debug and release builds. The returned UUID structure and public method signatures are identical, making this PR backwards-compatible.

## Benchmark

- **Godot version:** 2.1.6 (official Linux 64-bit binary)
- **CPU:** i7-6700K @ 4.4 GHz
- **OS:** Fedora 32

For 500,000 iterations:

- Debug (old): 18,200 ms
- **Debug (new):** 4,231 ms
- Release (old): 17,623 ms
- **Release (new):** 4,264 ms

### Benchmark code

```gdscript
extends Node2D

const uuid_util = preload("res://uuid_new.gd")

func _ready():
	var begin = OS.get_ticks_msec()
	var i = 0
	while i < 500000:
		uuid_util.v4()
		i += 1
	var end = OS.get_ticks_msec()
	print(end - begin)
	# Check if the returned UUID looks correct:
	print(uuid_util.v4())
```